refactor: moved Dashboard/Widgets component funcs to helpers file
